getting ready for release
skiasharp updated
splash screen text fix
dragging over band stack entries on the spectrum fix
added info about AI images on about box
high frequency spots fixed, caused by long to int truncation

+++ Project Files/Source/Console/display.cs
@@ -9536,7 +9536,7 @@ namespace Thetis
 
             if (rx == 1)
             {
-                vfo_hz = (int)vfoa_hz;
+                vfo_hz = vfoa_hz;
                 if (local_mox)
                 {
                     if (duplex)
@@ -9561,7 +9561,7 @@ namespace Thetis
             }
             else// rx == 2
             {
-                vfo_hz = (int)vfob_hz;
+                vfo_hz = vfob_hz;
                 if (local_mox)
                 {
                     if (duplex)
